+++ src/components/button/Button.js
@@ -8,6 +8,7 @@ export class ButtonComponent extends BaseComponent {
     let info = super.elementInfo();
     info.type = 'button';
     info.attr.type = (this.component.action === 'submit') ? 'submit' : 'button';
+    this.component.theme = this.component.theme || 'default';
     info.attr.class = 'btn btn-' + this.component.theme;
     if (this.component.block) {
       info.attr.class += ' btn-block';
@@ -27,7 +28,19 @@ export class ButtonComponent extends BaseComponent {
     this.setDisabled(this.button, disabled);
   }
 
+  getValue() {
+    if (!this.component.input) {
+      return;
+    }
+    return this.clicked;
+  }
+
   build() {
+    if (this.viewOnlyMode()) {
+      this.component.hidden = true;
+    }
+
+    this.clicked = false;
     this.createElement();
     this.element.appendChild(this.button = this.ce(this.info.type, this.info.attr));
     this.addShortcut(this.button);
@@ -54,6 +67,7 @@ export class ButtonComponent extends BaseComponent {
       }, true);
     }
     this.addEventListener(this.button, 'click', (event) => {
+      this.clicked = false;
       switch (this.component.action) {
         case 'submit':
           event.preventDefault();