+++ src/contrib/index.js
@@ -1,0 +1,6 @@
+import { StripeCheckoutComponent } from './stripe/checkout/StripeCheckout';
+module.exports = {
+  stripe: {
+    checkout: StripeCheckoutComponent
+  }
+};