+++ test/forms/simple.js
@@ -15,7 +15,8 @@ module.exports = {
           email: 'test@example.com',
           password: '123test',
           verifyPassword: '123test',
-          date: Harness.getDate()
+          date: Harness.getDate(),
+          submit: false
         }
       });
       done();
@@ -26,6 +27,7 @@ module.exports = {
         firstName: 'test',
         lastName: 'test2',
         email: 'bademail',
+        submit: false,
         password: '',
         verifyPassword: ''
       }}, [{