+++ Changelog.md
@@ -4,12 +4,64 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Added
+ - CSS class ('radio-selected') for selected option of Radio component
+
+### Fixed
+ - Select Resource component searching.
+
+## 2.27.0
+### Fixed
+ - Issue where read-only forms would still try to submit.
+ - Problem with read-only wizards triggering beforeSubmit handlers.
+ - Fix issue where submissions made before revisions are made will sometimes cause the form to not load.
+
+### Changed
+ - Upgrade all dependencies
+
+## 2.26.2
+### Fixed
+ - Problem where a component has input should also return true if it has inputs.
+
+## 2.26.1
+### Fixed
+ - Problems where data keys are added even if component is not set with input.
+ - Failing tests.
+ - Datagrid data merging.
+
+### Changed
+ - Upgraded choices.js to 3.0.3 which includes performance fix.
+ - Removed performance hack in Select since 3.0.3 of choices resolves the problem.
+
+## 2.26.0
+### Added
+ - New contributed module system with Stripe integration.
+ - A way to pass the formio instance object to the currentUser and accessInfo methods.
+
+## 2.25.8
+### Added
+ - Support for JSONLogic dates.
+ - Added 'searchEnabled' option and defaulted it to false, user can enable it with component property
+ - Updated default value for 'removeItemButton' option to multi-select OR false, and if needed user can enable it with the component property
+ - Date formatting based on the locale configuration
+ - viewAsHTML feature.
+ - Confirmation dialog before a form/wizard is canceled.
+
+## 2.25.7
+### Fixed
+ - Problem with the Select dropdown from re-rendering after it has been destroyed.
+ - Issue with the default select for html5 widgets.
+
+### Changed
+ - Upgrade moment to 2.20.1
+
 ## 2.25.6
 ### Added
  - Ability to dynamically alter text based on input data.
  - Try/catch around the jsLogic for checkconditionals.
  - LazyLoading for select dropdowns.
- 
+
 ### Fixed
  - Caching issue with getTempToken method.
  - Problem with the "in" operator for JSONLogic crashing with null inputs.
@@ -42,10 +94,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - The ability to render Select component as plain select dropdown using widget: 'html5' setting.
  - Performance improvements to language selection
- 
+
 ### Changed
  - Moved all translation capabilities into FormioForm for performance reasons.
- 
+
 ### Fixed
  - Issues with Lodash operators for JSONLogic.
 
@@ -87,7 +139,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
  - File component inside Datagrid component.
  - Components with label position inside Datagrid component.
- 
+
 ### Added
  - Interpolation to the select headers when requests are made.
  - Option to make the wizard header buttons not clickable.